feat(product page): Implement add to cart bar
