+++ frontend/components/ui/Rating.tsx
@@ -1,15 +1,16 @@
-import { HStack, StackProps } from '@chakra-ui/react';
+import { HStack, ResponsiveValue, StackProps } from '@chakra-ui/react';
 import { faStarHalf } from '@fortawesome/pro-duotone-svg-icons';
 import { faStar } from '@fortawesome/pro-solid-svg-icons';
 import { FaIcon, FaIconProps } from '@ifixit/icons';
 
 export interface RatingProps extends StackProps {
    value?: number;
+   size?: ResponsiveValue<string | number>;
 }
 
 const stars = new Array(5).fill(null).map((_, index) => index + 1);
 
-export const Rating = (props: RatingProps) => {
+export function Rating(props: RatingProps) {
    const { value = 5, ...rest } = props;
 
    const halfStarsValue = Math.round(value * 2);
@@ -23,14 +24,17 @@ export const Rating = (props: RatingProps) => {
             } else if (halfStarsValue >= i * 2 - 1) {
                appearance = RatingStarAppearance.Half;
             }
-            return <RatingStar key={i} appearence={appearance} />;
+            return (
+               <RatingStar key={i} size={props.size} appearence={appearance} />
+            );
          })}
       </HStack>
    );
-};
+}
 
 type RatingStarProps = Omit<FaIconProps, 'icon'> & {
    appearence: RatingStarAppearance;
+   size?: ResponsiveValue<string | number>;
 };
 
 export enum RatingStarAppearance {
@@ -41,22 +45,25 @@ export enum RatingStarAppearance {
 
 export const RatingStar = ({
    appearence = RatingStarAppearance.Empty,
+   size = '4',
    ...otherProps
 }: RatingStarProps) => {
    switch (appearence) {
       case RatingStarAppearance.Empty: {
-         return <FaIcon icon={faStar} h="4" color="gray.300" {...otherProps} />;
+         return (
+            <FaIcon icon={faStar} h={size} color="gray.300" {...otherProps} />
+         );
       }
       case RatingStarAppearance.Full: {
          return (
-            <FaIcon icon={faStar} h="4" color="brand.500" {...otherProps} />
+            <FaIcon icon={faStar} h={size} color="brand.500" {...otherProps} />
          );
       }
       case RatingStarAppearance.Half: {
          return (
             <FaIcon
                icon={faStarHalf}
-               h="4"
+               h={size}
                sx={{
                   '--fa-primary-color': 'var(--chakra-colors-brand-500)',
                   '--fa-secondary-color': 'var(--chakra-colors-gray-300)',
